add new tags into database
