+++ README.md
@@ -1,2 +1,9 @@
 # blog-rs
-Blog written in Rust (Warp + Yew)
+A singleton self-serve Blog written in Rust (Warp + Yew)
+
+Current version: `0.2.0`
+
+## Features
+1. Self-serve
+1. Embed `Markdown` editor
+1. Export posts for `Hugo`
