+++ backend/src/db/tag.rs
@@ -9,6 +9,7 @@ use std::{
 };
 
 use ahash::AHasher;
+use blog_common::result::Error;
 use bytes::{Buf, Bytes, BytesMut};
 use lazy_static::lazy_static;
 use parking_lot::RwLock;
@@ -17,8 +18,6 @@ use tokio::{
     fs::{remove_file, rename, File, OpenOptions},
     io::{self, AsyncReadExt, AsyncSeekExt, AsyncWriteExt, BufReader, BufWriter},
 };
-
-use blog_common::result::Error;
 
 use crate::{
     db::{self, model::Tag, DATA_SOURCE},
@@ -55,7 +54,7 @@ pub async fn get_names(id_array: Vec<i64>) -> Result<Vec<String>> {
 pub(super) async fn record_usage(post_id: u64, tags: &Vec<String>) -> Result<()> {
     // query id list by name list
     let mut sql = String::with_capacity(256);
-    sql.push_str("SELECT id from tag WHERE name IN (");
+    sql.push_str("SELECT id,name from tag WHERE name IN (");
     for _i in 0..tags.len() {
         sql.push_str("?,");
     }
@@ -64,10 +63,33 @@ pub(super) async fn record_usage(post_id: u64, tags: &Vec<String>) -> Result<()>
     for tag in tags.iter() {
         query = query.bind(tag);
     }
-    let tags = query.fetch_all(&DATA_SOURCE.get().unwrap().sqlite).await?;
+    let mut tags_in_db = query.fetch_all(&DATA_SOURCE.get().unwrap().sqlite).await?;
+
+    // 查看有没有新的tag
+    if tags_in_db.len() < tags.len() {
+        let mut new_tags: Vec<Tag> = Vec::with_capacity(tags.len() - tags_in_db.len());
+        {
+            let mut tags_in_db_iter = tags_in_db.iter();
+            for tag in tags.iter() {
+                if !tags_in_db_iter.any(|e| e.name.eq(tag)) {
+                    let id = sqlx::query("INSERT INTO tag(name, created_at)VALUES(?,NOW())")
+                        .bind(tag)
+                        .execute(&DATA_SOURCE.get().unwrap().sqlite)
+                        .await?
+                        .last_insert_rowid();
+                    let new_tag = Tag {
+                        id,
+                        name: String::from(tag),
+                    };
+                    new_tags.push(new_tag);
+                }
+            }
+        }
+        tags_in_db.append(&mut new_tags);
+    }
 
     let post_id = post_id as i64;
-    for tag in tags {
+    for tag in tags_in_db {
         sqlx::query("INSERT INTO tag_usage(post_id, tag_id)VALUES(?,?)")
             .bind(post_id)
             .bind(tag.id)
