+++ backend/src/service/asset_list.rs
@@ -1,7 +1,10 @@
 [
-("index-665b7454336b637.js", include_bytes!(r#"..\resource\asset/index-665b7454336b637.js.gz"#), "text/javascript"),
-("index-665b7454336b637_bg.wasm", include_bytes!(r#"..\resource\asset/index-665b7454336b637_bg.wasm.gz"#), "application/wasm"),
+("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
+("index-5114c039e9a0540e.js", include_bytes!(r#"..\resource\asset/index-5114c039e9a0540e.js.gz"#), "text/javascript"),
+("index-5114c039e9a0540e_bg.wasm", include_bytes!(r#"..\resource\asset/index-5114c039e9a0540e_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
 ("purecss.min-b565cdb90b7aac9f.css", include_bytes!(r#"..\resource\asset/purecss.min-b565cdb90b7aac9f.css.gz"#), "text/css"),
 ("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
+("toastui-editor-all.min.js", include_bytes!(r#"..\resource\asset/toastui-editor-all.min.js.gz"#), "text/javascript"),
+("toastui-editor.min.css", include_bytes!(r#"..\resource\asset/toastui-editor.min.css.gz"#), "text/css"),
 ]
